fix(route): rename nddaily to oeeee
