+++ docs/traditional-media.md
@@ -1662,7 +1662,7 @@ category 对应的关键词有
 
 ### 奥一网
 
-<Route author="TimWu007" example="/nddaily/oeeee/170" path="/nddaily/oeeee/:channel" :paramsDesc="['频道 ID']">
+<Route author="TimWu007" example="/oeeee/web/170" path="/oeeee/web/:channel" :paramsDesc="['频道 ID']">
 
 -   若在桌面端打开奥一网栏目页（如 <https://www.oeeee.com/api/channel.php?s=/index/index/channel/gz），可查看该页源代码，搜索> `OECID`。
 -   若在移动端打开奥一网栏目页（格式例：<https://m.oeeee.com/m.php?s=/m2/channel&channel_id=169），即可从> url 中获取。需注意的是，如果该栏目页的 url 格式为 <https://m.oeeee.com/detailChannel_indexData.html?channel_id=266> ，则 `266` 并非为本路由可用的频道 ID，建议从桌面端获取。
@@ -1671,7 +1671,7 @@ category 对应的关键词有
 
 ### 南都客户端（按南都号 ID）
 
-<Route author="TimWu007" example="/nddaily/ndapp/channel/50" path="/nddaily/ndapp/channel/:id?" :paramsDesc="['南都号 ID']">
+<Route author="TimWu007" example="/oeeee/app/channel/50" path="/oeeee/app/channel/:id?" :paramsDesc="['南都号 ID']">
 
 南都号的 UID 可通过 `m.mp.oeeee.com` 下的文章页面获取。点击文章上方的南都号头像，进入该南都号的个人主页，即可从 url 中获取。
 
@@ -1679,7 +1679,7 @@ category 对应的关键词有
 
 ### 南都客户端（按记者）
 
-<Route author="TimWu007" example="/nddaily/ndapp/reporter/249" path="/nddaily/ndapp/reporter/:id" :paramsDesc="['记者 UID']">
+<Route author="TimWu007" example="/oeeee/app/reporter/249" path="/oeeee/app/reporter/:id" :paramsDesc="['记者 UID']">
 
 记者的 UID 可通过 `m.mp.oeeee.com` 下的文章页面获取。点击文章下方的作者头像，进入该作者的个人主页，即可从 url 中获取。
