+++ lib/v2/oeeee/maintainer.js
@@ -1,3 +1,5 @@
 module.exports = {
-    '/:channel': ['TimWu007'],
+    '/web/:channel': ['TimWu007'],
+    '/app/channel/:id': ['TimWu007'],
+    '/app/reporter/:id': ['TimWu007'],
 };
