+++ lib/v2/oeeee/router.js
@@ -1,5 +1,5 @@
 module.exports = (router) => {
-    router.get('/oeeee/:channel', require('./oeeee'));
-    router.get('/ndapp/channel/:id', require('./ndapp/channel'));
-    router.get('/ndapp/reporter/:id', require('./ndapp/reporter'));
+    router.get('/web/:channel', require('./web'));
+    router.get('/app/channel/:id', require('./app/channel'));
+    router.get('/app/reporter/:id', require('./app/reporter'));
 };
